Move binary search to general utils
